+++ ConvertPG2SS/Program.cs
@@ -63,7 +63,7 @@ namespace ConvertPG2SS {
 			_param.WriteParametersToLog();
 #endif
 			if (General.CheckParams(_param, _log)) {
-				if (bool.Parse(_param["other.pg2ss"].ToString())) 
+				if (bool.Parse(_param[Parameters.OtherPg2Ss].ToString())) 
 					ProcessPgSchema.Do();
 					ProcessPgBulk.Do();				
 			}